renamed the fuzzer attack file
