+++ pyrit/attacks/__init__.py
@@ -20,7 +20,7 @@ from pyrit.attacks.multi_turn.tree_of_attacks import (
     TAPAttackResult,
 )
 from pyrit.attacks.single_turn.prompt_sending import PromptSendingAttack
-from pyrit.attacks.fuzzer_attack import FuzzerAttack, FuzzerAttackContext, FuzzerAttackResult
+from pyrit.attacks.fuzzer import FuzzerAttack, FuzzerAttackContext, FuzzerAttackResult
 
 __all__ = [
     "AttackAdversarialConfig",
